Add of SLAM and goal to one robot
This solves issue #4
When publishing the tf of the robot, each link will now have a link corresponding to the robot namespace.

This starts issue #6
This commit enable robot_0 to do SLAM and go to one goal.
AMCL, map_server and move_base have been added to ptdr_spawn.

The same process needs to be be applied to robot_1 (with a cleaner code if possible).