Spawning function
This improve issue #5.

The ptdr_spawn.launch is now a place to declare how many robots are going to spawn and with which parameters. It will then call the spawn function.

This commit improve readability for spawn and the SLAM features associated to the robot.

An rviz config files has been staged as well and needs improvement.