Exploration of the map thansk to rrt
This update issue #10.

This commit adds a launch file to launch after the spawning and map merging one. After setting the boundaries of the world and a rough localisation of the robot on the map on Rviz , it will trigger the exploration of the map.

The issue is not solved yet. robot_1 and robot_2 moves is erratic after few minutes and robot_0 is not moving. Due to map_merging #8 ?