Exploration of the map improvement
This commit update the issue #7 #9 #10 #11

The map is merged in the map merged topic which helps in readability.

Frequency has been increased. Extensive tests have to be made.

Autonomous exploration of the map seems to be acheived. Extensive tests have to be made to close the issue.

Robots behavior is no more erratic closing the issue.